+++ packages/antwerp-ui/react-components/src/utils/file.utils.ts
@@ -37,6 +37,5 @@ export function invalidIcon(icon: SVGGraphicsElement): boolean {
 export function iconExists(iconName: string): boolean {
   const svgContainer = document.getElementById(ICONS_SVG_HTML_ID);
   if (!svgContainer) return false;
-
   return !!svgContainer.querySelector(`symbol#ai-${iconName}`);
 }